+++ ffi_common/src/datetime.rs
@@ -2,24 +2,18 @@
 //! FFI support for exposing time stamps.
 //!
 
-use crate::{declare_value_type_array_struct, error};
+use crate::declare_value_type_array_struct;
 use chrono::NaiveDateTime;
 use paste::paste;
 
 /// Represents a UTC timestamp in a way that's safe to transfer across the FFI boundary.
 #[repr(C)]
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, Default)]
 pub struct TimeStamp {
     /// Seconds since the UNIX epoch time (January 1, 1970).
     secs: i64,
     /// Nanoseconds since the last whole second.
     nsecs: u32,
-}
-
-impl Default for TimeStamp {
-    fn default() -> Self {
-        Self { secs: 0, nsecs: 0 }
-    }
 }
 
 declare_value_type_array_struct!(TimeStamp);
@@ -36,12 +30,6 @@ impl From<&NaiveDateTime> for TimeStamp {
     }
 }
 
-impl From<TimeStamp> for NaiveDateTime {
-    fn from(timestamp: TimeStamp) -> Self {
-        Self::from_timestamp(timestamp.secs, timestamp.nsecs)
-    }
-}
-
 impl From<&TimeStamp> for NaiveDateTime {
     fn from(timestamp: &TimeStamp) -> Self {
         Self::from_timestamp(timestamp.secs, timestamp.nsecs)
@@ -49,13 +37,14 @@ impl From<&TimeStamp> for NaiveDateTime {
 }
 
 // Option conversion impls
-impl From<&Option<NaiveDateTime>> for OptionTimeStamp {
-    fn from(opt: &Option<NaiveDateTime>) -> Self {
-        let d = opt.map(|s| TimeStamp {
+impl From<Option<&NaiveDateTime>> for OptionTimeStamp {
+    fn from(opt: Option<&NaiveDateTime>) -> Self {
+        opt.map(|s| TimeStamp {
             secs: s.timestamp(),
             nsecs: s.timestamp_subsec_nanos(),
-        });
-        (&d).into()
+        })
+        .as_ref()
+        .into()
     }
 }
 
@@ -73,10 +62,10 @@ impl From<OptionTimeStamp> for Option<NaiveDateTime> {
 }
 
 // Collection conversion impls
-impl From<&Vec<NaiveDateTime>> for FFIArrayTimeStamp {
-    fn from(v: &Vec<NaiveDateTime>) -> Self {
-        let timestamps: Vec<TimeStamp> = v.iter().map(|e| e.into()).collect();
-        (&timestamps).into()
+impl From<&[NaiveDateTime]> for FFIArrayTimeStamp {
+    fn from(slice: &[NaiveDateTime]) -> Self {
+        let timestamps: Vec<TimeStamp> = slice.iter().map(|e| e.into()).collect();
+        timestamps.as_slice().into()
     }
 }
 
@@ -86,6 +75,29 @@ impl From<FFIArrayTimeStamp> for Vec<NaiveDateTime> {
         unsafe {
             let v = Vec::from_raw_parts(array.ptr as *mut TimeStamp, array.len, array.cap);
             v.iter().map(|e| e.into()).collect()
+        }
+    }
+}
+
+impl From<Option<&[NaiveDateTime]>> for FFIArrayTimeStamp {
+    fn from(slice: Option<&[NaiveDateTime]>) -> Self {
+        slice.map_or(
+            Self {
+                ptr: std::ptr::null(),
+                len: 0,
+                cap: 0,
+            },
+            |s| s.into(),
+        )
+    }
+}
+
+impl From<FFIArrayTimeStamp> for Option<Vec<NaiveDateTime>> {
+    fn from(array: FFIArrayTimeStamp) -> Self {
+        if array.ptr.is_null() {
+            None
+        } else {
+            Some(array.into())
         }
     }
 }