+++ ffi_derive/src/enum_ffi.rs
@@ -4,16 +4,13 @@
 
 use heck::SnakeCase;
 use proc_macro2::TokenStream;
-use quote::quote;
+use quote::{format_ident, quote};
 use syn::Ident;
 
 /// Builds an FFI module for the enum `type_name`.
 ///
 pub(super) fn build(module_name: &Ident, type_name: &Ident) -> TokenStream {
-    let fn_name = Ident::new(
-        &["free_", &type_name.to_string().to_snake_case()].concat(),
-        type_name.span(),
-    );
+    let fn_name = format_ident!("free_{}", &type_name.to_string().to_snake_case());
 
     quote! {
         #[allow(missing_docs)]
@@ -25,7 +22,6 @@ pub(super) fn build(module_name: &Ident, type_name: &Ident) -> TokenStream {
 
             #[no_mangle]
             pub unsafe extern "C" fn #fn_name(data: #type_name) {
-                ffi_common::error::clear_last_err_msg();
                 let _ = data;
             }